refactor: consolidate special buttons
